requirements.txt
anyio==4.7.0
base58==2.1.1
certifi==2024.12.14
charset-normalizer==3.4.1
Deprecated==1.2.15
ECPy==1.2.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycryptodome==3.21.0
requests==2.32.3
sniffio==1.3.1
types-Deprecated==1.2.15.20241117
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
wrapt==1.17.0
xrpl-py==4.0.0
